gear refactored
